Add machine learning trade prediction and index creation
This commit adds the functionality for machine learning based trade prediction. In addition, it includes the creation of a unique index on 'pumpfun_trades' table. File structures have also been revised as part of this update, such as renaming 'request.go' to 'simple.go' and moving it into a new subdirectory 'requests'.